fix: removed reset-buttom from some unnecessary pages
